+++ ui/timeline/src/commonMain/kotlin/social/androiddev/timeline/HorizontalSpacer.kt
@@ -1,0 +1,12 @@
+package social.androiddev.timeline
+
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.width
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun HorizontalSpacer() {
+    Spacer(Modifier.width(12.dp))
+}
