+++ ui/timeline/src/commonMain/kotlin/social/androiddev/timeline/VerticalSpacer.kt
@@ -1,0 +1,12 @@
+package social.androiddev.timeline
+
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun VerticalSpacer() {
+    Spacer(Modifier.height(8.dp))
+}
